ETLP-6: adds Python package layout
