+++ src/attendance_etl/__init__.py
@@ -1,0 +1,1 @@
+"""Attendance ETL package."""
